Optimize TxQuery
